Bear no longer leaks
