--- src/components/bjorn/BjornBildeLiten.tsx
@@ -1,122 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import BjornBildeStor from "./BjornBildeStor";
-
-class BjornBildeLiten extends BjornBildeStor {
-  render() {
-    /* eslint-disable */
-    return (
-      <svg
-        width="64px"
-        height="64px"
-        viewBox="0 0 64 64"
-        version="1.1"
-        xmlns="http://www.w3.org/2000/svg"
-        xmlnsXlink="http://www.w3.org/1999/xlink"
-      >
-        <title>{this.props.alt}</title>
-        <defs>
-          <circle id="path-1-bjorn" cx="32" cy="32" r="32" />
-        </defs>
-        <g
-          id="Page-1"
-          stroke="none"
-          strokeWidth="1"
-          fill="none"
-          fillRule="evenodd"
-        >
-          <g id="Group-5">
-            <g id="Group-3">
-              <g id="Group-10-Copy-3">
-                <g id="Group-6">
-                  <mask id="mask-2-bjorn" fill="white">
-                    <use xlinkHref="#path-1-bjorn" />
-                  </mask>
-                  <use id="Mask" fill="#A4DFF7" xlinkHref="#path-1-bjorn" />
-                  <g id="Group-150" mask="url(#mask-2-bjorn)">
-                    <g transform="translate(6.400000, 3.840000)">
-                      <g id="Group" transform="translate(0.324907, 0.170667)">
-                        <path
-                          d="M25.1245478,45.154717 L17.5391463,45.154717 C13.4494717,45.154717 0,53.5789644 0,61.4209844 L0,82.0553459 L50.2490956,82.0553459 L50.2490956,61.4209844 C50.2490956,53.5789644 36.7996239,45.154717 32.7099493,45.154717 L25.1245478,45.154717 Z"
-                          id="Fill-29-Copy"
-                          fill="#624B7F"
-                        />
-                        <path
-                          d="M45.6821681,20.5040209 C45.4722966,20.5040209 45.2745511,20.5467175 45.0907969,20.6172597 C44.9779327,20.6302543 44.8715978,20.6367516 44.7596664,20.6488181 C43.032191,10.0025138 35.0178992,1.94213836 25.3684755,1.94213836 C15.8515039,1.94213836 7.92769003,9.78531805 6.0565693,20.2144266 C5.99780529,20.310958 5.92691537,20.4158431 5.88214279,20.5040209 C5.88214279,20.5040209 5.77953896,20.7110066 5.64708674,20.6172597 C5.46333261,20.5467175 5.26558705,20.5040209 5.05664835,20.5040209 C4.14907084,20.5040209 3.41498708,21.234504 3.41498708,22.1367017 L3.41498708,28.7268291 C3.41498708,29.6290268 4.14907084,30.3595099 5.05664835,30.3595099 C5.4194928,30.3595099 5.75062333,30.2369892 6.0220571,30.0402136 C7.46970385,36.6266281 11.3443975,42.0704446 16.4633959,45.0480683 C15.0698493,45.6049805 12.7286165,46.8394691 12.7286165,46.8394691 C13.7583858,48.3459165 16.6853932,55.3509434 25.3638116,55.3509434 C34.0412973,55.3509434 36.9766996,48.3459165 38.0074017,46.8394691 C38.0074017,46.8394691 35.9926356,45.7386394 34.3714951,44.9914489 C39.4419898,42.0017587 43.2756419,36.5848597 44.713961,30.0402136 C44.9881931,30.2369892 45.3183909,30.3595099 45.6821681,30.3595099 C46.5878801,30.3595099 47.3219638,29.6290268 47.3219638,28.7268291 L47.3219638,22.1367017 C47.3219638,21.234504 46.5878801,20.5040209 45.6821681,20.5040209"
-                          id="Fill-44-Copy"
-                          fill="#E7E5E2"
-                        />
-                        <g
-                          id="blid-mann"
-                          transform="translate(16.489509, 21.363522)"
-                        >
-                          <path
-                            d="M1.49174441,4.77918496 C0.150483845,4.87973917 -0.224920806,2.91201313 0.178291597,1.62694873 C0.255226377,1.38340458 0.699223483,0.275463195 1.48525594,0.275463195 C2.27036147,0.275463195 2.61703144,0.880633509 2.66523155,0.984877786 C3.24270586,2.23765414 2.95906679,4.66848307 1.49174441,4.77918496"
-                            id="Fill-61"
-                            fill={this.state.oyefarge}
-                          />
-                          <path
-                            d="M16.4610281,4.77918496 C17.8022886,4.87973917 18.1776933,2.91201313 17.7735539,1.62694873 C17.6975461,1.38340458 17.253549,0.275463195 16.4675165,0.275463195 C15.682411,0.275463195 15.335741,0.880633509 15.2875409,0.984877786 C14.7100666,2.23765414 14.9927787,4.66848307 16.4610281,4.77918496"
-                            id="Fill-63"
-                            fill={this.state.oyefarge}
-                          />
-                          <path
-                            d={this.state.munnform}
-                            id="Stroke-67"
-                            stroke="#D19E9C"
-                            strokeWidth="1"
-                            strokeLinecap="round"
-                          />
-                        </g>
-                        <path
-                          d="M7.36949754,26.7045692 L7.36949754,24.9508146 C7.23971079,20.0701706 8.9902492,17.8040832 10.4316208,16.0824791 C10.4316208,16.0824791 8.95859389,20.1635106 9.56848612,19.3836032 C13.0168041,14.9810673 29.4680666,19.2772993 32.7823771,13.5949686 C33.1981168,16.7228951 30.9252658,16.0824791 30.9252658,16.0824791 C34.2596247,16.6933374 35.3580638,15.3668726 36.141005,13.8257257 C36.3995234,15.0827042 36.7266282,14.8306862 35.568044,15.6914883 C39.6652958,14.94062 39.2822666,13.8257257 39.2822666,13.8257257 C39.8647242,17.4442057 43.3964011,15.5276248 43.3964011,21.4536766 L43.3964011,26.7045692 C43.3964011,26.9731809 43.6179883,27.1899371 43.8912791,27.1899371 L44.3808812,27.1899371 C44.654172,27.1899371 44.8768143,26.9731809 44.8768143,26.7045692 L44.8768143,22.803995 C45.1300567,9.14005858 37.166637,0 26.2054597,0 C15.7423259,0 5.85426357,9.11205658 5.85953945,21.1757309 C5.858836,21.1826449 5.85707737,23.025591 5.85426357,26.7045692 C5.85426357,26.9731809 6.07479553,27.1899371 6.34808634,27.1899371 L6.87461959,27.1899371 C7.02550988,27.1899371 7.15635181,27.1194136 7.24498667,27.0157025 C7.33889741,26.8715441 7.36949754,26.7927236 7.36949754,26.7045692 Z"
-                          id="Fill-65-Copy"
-                          fill="#7F756C"
-                        />
-                        <path
-                          d="M25.7999489,28.2256424 C26.611425,28.0983534 27.4022624,28.1289793 27.773759,28.5941103 C28.6790474,29.7282261 28.4191874,31.0384408 26.894738,32.0069852 C26.1329824,32.4903004 25.0541413,32.6884117 24.3927649,32.3898091"
-                          id="Stroke-65-Copy"
-                          stroke="#D1BFA3"
-                          strokeWidth="1"
-                          strokeLinecap="round"
-                        />
-                      </g>
-                      <g
-                        id="Group-149"
-                        transform="translate(5.871573, 30.037333)"
-                      >
-                        <path
-                          d="M0.417539191,0.405729167 C1.05990718,1.90518353 2.03119076,4.36697989 2.21133045,4.89796627 C2.47098709,5.66333975 2.98312611,6.89831619 3.74774752,8.60289559 L4.6795132,10.0320188 L1.6439212,10.7360367 C1.90569934,9.67109606 2.06850531,8.85948065 2.13233911,8.30119048 C2.16232841,8.03890416 1.87561781,7.46443921 1.6439212,6.33239087 C1.48225994,5.54252877 1.07336643,3.07020942 0.121736799,0.525068204 C0.0854678568,0.428066623 0.144950187,0.431134469 0.223263201,0.405729167 C0.305863439,0.378933058 0.334469194,0.211822134 0.417539191,0.405729167 Z"
-                          id="Rectangle-3"
-                          fill="#7F756C"
-                        />
-                        <path
-                          d="M33.9747525,0.39936756 C34.0715439,0.707017144 34.3854139,1.48976971 34.9163624,2.74762525 C35.447311,4.00548079 36.2457244,5.857808 37.3116027,8.30460689 L38.9936675,10.6858507 L35.7135107,10.3701265 C35.9464522,9.86643381 36.0629229,9.31454475 36.0629229,8.71445933 C36.0629229,8.21716832 35.4920203,6.22760029 34.5359942,2.74762525 C34.4547293,2.45181754 34.0981248,0.925891738 33.9747525,0.39936756 C33.9236606,0.181319346 33.9236606,0.181319346 33.9747525,0.39936756 Z"
-                          id="Rectangle-3"
-                          fill="#7F756C"
-                          transform="translate(36.465051, 5.460841) scale(-1, 1) translate(-36.465051, -5.460841) "
-                        />
-                        <polygon
-                          id="Rectangle-3"
-                          fill="#7F756C"
-                          points="17.9566997 12.0634921 21.2366192 12.0634921 20.8757024 15.9238095 18.3579425 15.9238095"
-                        />
-                      </g>
-                    </g>
-                  </g>
-                </g>
-              </g>
-            </g>
-          </g>
-        </g>
-      </svg>
-    );
-    /* eslint-enable */
-  }
-}
-
-BjornBildeLiten.propTypes = {
-  alt: PropTypes.string,
-};
-
-export default BjornBildeLiten;
